Move relevant structs & enums to beambase
